refactor: rename kuzu/ → ladybug/, Kuzu*Operations → Ladybug*Operations, add GraphProvider.LADYBUG
- git mv graphiti_core/driver/kuzu/ → graphiti_core/driver/ladybug/
- Rename all 11 Kuzu*Operations classes to Ladybug*Operations
- Rename parse_kuzu_entity_node/edge → parse_ladybug_entity_node/edge
- Replace GraphProvider.KUZU → GraphProvider.LADYBUG in enum
- Update ladybug_driver.py imports and provider references
- Delete kuzu_driver.py
- Rename test_kuzu_hnsw_search.py → test_ladybug_hnsw_search.py
