+++ graphiti_core/driver/ladybug/operations/community_edge_ops.py
@@ -41,14 +41,14 @@ def _community_edge_from_record(record: Any) -> CommunityEdge:
     )
 
 
-class KuzuCommunityEdgeOperations(CommunityEdgeOperations):
+class LadybugCommunityEdgeOperations(CommunityEdgeOperations):
     async def save(
         self,
         executor: QueryExecutor,
         edge: CommunityEdge,
         tx: Transaction | None = None,
     ) -> None:
-        query = get_community_edge_save_query(GraphProvider.KUZU)
+        query = get_community_edge_save_query(GraphProvider.LADYBUG)
         params: dict[str, Any] = {
             'community_uuid': edge.source_node_uuid,
             'entity_uuid': edge.target_node_uuid,
