+++ graphiti_core/driver/ladybug/operations/episodic_edge_ops.py
@@ -42,7 +42,7 @@ def _episodic_edge_from_record(record: Any) -> EpisodicEdge:
     )
 
 
-class KuzuEpisodicEdgeOperations(EpisodicEdgeOperations):
+class LadybugEpisodicEdgeOperations(EpisodicEdgeOperations):
     async def save(
         self,
         executor: QueryExecutor,
@@ -70,8 +70,8 @@ class KuzuEpisodicEdgeOperations(EpisodicEdgeOperations):
         tx: Transaction | None = None,
         batch_size: int = 100,
     ) -> None:
-        # Kuzu doesn't support UNWIND - iterate and save individually
-        query = get_episodic_edge_save_bulk_query(GraphProvider.KUZU)
+        # LadybugDB doesn't support UNWIND - iterate and save individually
+        query = get_episodic_edge_save_bulk_query(GraphProvider.LADYBUG)
         for edge in edges:
             params: dict[str, Any] = {
                 'source_node_uuid': edge.source_node_uuid,
